refactor(ui): extract shared UI formatting helpers
